% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrlda.R
\name{discriminant_fun}
\alias{discriminant_fun}
\title{Discriminant function which measures the distance between a point and a centroid j using Mahalanobis distance}
\usage{
discriminant_fun(x, centroid, prior)
}
\arguments{
\item{x}{[1,r] data point that has dimensionality up to the rth subspace}

\item{centroid}{[1,r] centroid point j that has dimensionality up to the rth subspace}

\item{prior}{[1] prior probability for class j}
}
\value{
Mahalanobis distance between x and centroid j
}
\description{
Discriminant function which measures the distance between a point and a centroid j using Mahalanobis distance
}
