% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrlda.R
\name{fs.classify.lda}
\alias{fs.classify.lda}
\title{Linear Discriminant Analysis (LDA) Prediction}
\usage{
fs.classify.lda(X, Y, X.test, Y.test = NULL)
}
\arguments{
\item{X}{[n1, r] data matrix with n samples in r dimensions for training.}

\item{Y}{[n1] the labels for each sample for training.}

\item{X.test}{[n2, r] data matrix with n samples in r dimensions for testing.}

\item{Y.test}{[n2] an optional argument for the labels for each sample for testing.}

\item{ylabs}{[K] vector containing the unique, ordered class labels.}

\item{centroids}{[K, C-1] centroid matrix of the unique, ordered class labels.}

\item{priors}{[K] vector containing prior probability for the unique, ordered class labels.}
}
\value{
Xr [n, K-1] projected data matrix.

Mp [K, K-1] projected centroid matrix.

Yhat [n] prediction matrix containing predictions for each example.
}
\description{
A function for using Linear Discriminant Analysis (LDA) for prediction.
}
\author{
Richard Chen, modified by Eric Bridgeford
}
