% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrlda.R
\name{fs.predict.lda}
\alias{fs.predict.lda}
\title{Linear Discriminant Analysis (LDA) Prediction}
\usage{
fs.predict.lda(X, ylabs, centroids, priors, A)
}
\arguments{
\item{ylabs}{[C] vector containing the unique, ordered class labels.}

\item{centroids}{[C, r] centroid matrix of the unique, ordered class labels.}

\item{priors}{[C] vector containing prior probability for the unique, ordered class labels.}

\item{A}{[d, C-1] the projection matrix from d to C-1 dimensions.}

\item{Xr}{[n, r] data matrix with n samples in r dimensions.}
}
\value{
Xr [n, C-1] projected data matrix.

Mp [C, C-1] projected centroid matrix.

Yhat [n] prediction matrix containing predictions for each example.
}
\description{
A function for using Linear Discriminant Analysis (LDA) for prediction.
}
\author{
Richard Chen
}
