% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{fs.project.cpca}
\alias{fs.project.cpca}
\title{Class PCA}
\usage{
fs.project.cpca(X, Y, r)
}
\arguments{
\item{X}{[n, d] the data with n samples in d dimensions.}

\item{Y}{[n] the labels of the samples.}

\item{r}{the rank of the projection.}
}
\value{
A [d, r] the projection matrix from d to r dimensions.

ylabs [C] vector containing the unique, ordered class labels.

centroids [C, d] centroid matrix of the unique, ordered classes.

priors [C] vector containing prior probability for the unique, ordered classes.

Xr [n, r] the data in reduced dimensionality.

cr [C, r] the centroids in reduced dimensionality.
}
\description{
A function that performs PCA on the class-centered data.
}
\author{
Eric Bridgeford
}
