% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrlda.R
\name{fs.project.lda}
\alias{fs.project.lda}
\title{Linear Discriminant Analysis (LDA)}
\usage{
fs.project.lda(X, Y, r)
}
\arguments{
\item{X}{[n, d] the data with n samples in d dimensions.}

\item{Y}{[n] the labels of the samples.}
}
\value{
A [d, r] the projection matrix

ylabs [K] vector containing the unique, ordered class labels.

centroids [K, d] centroid matrix of the unique, ordered classes.

priors [K] vector containing prior probability for the unique, ordered classes.

Xr [n, r] the data in reduced dimensionality.

cr [K, r] the centroids in reduced dimensionality.
}
\description{
A function for implementing the LDA Algorithm.
}
\author{
Richard Chen, modified by Eric Bridgeford
}
