% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol.R
\name{fs.project.lol}
\alias{fs.project.lol}
\title{Linear Optimal Low-Rank Projection (LOL)}
\usage{
fs.project.lol(X, Y, r)
}
\arguments{
\item{X}{[n, d] the data with n samples in d dimensions.}

\item{Y}{[n] the labels of the samples.}

\item{r}{the rank of the projection.}
}
\value{
A [d, r] the projection matrix for the linearly optimal projection from d to r dimensions.
}
\description{
A function for implementing the Linear Optimal Low-Rank Projection (LOL) Algorithm.
}
\author{
Eric Bridgeford, adapted from Joshua Vogelstein
}
