% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrcca.R
\name{lol.project.lrcca}
\alias{lol.project.lrcca}
\title{Low-rank Canonical Correlation Analysis (LR-CCA)}
\usage{
lol.project.lrcca(X, Y, r)
}
\arguments{
\item{X}{[n, p] the data with n samples in d dimensions.}

\item{Y}{[n, q] the labels of the samples.}

\item{r}{the rank of the projection.}
}
\value{
A [d, r] the projection matrix from d to r dimensions.

ylabs [K] vector containing the unique, ordered class labels.

centroids [K, d] centroid matrix of the unique, ordered classes.

priors [K] vector containing prior probability for the unique, ordered classes.

Xr [n, r] the data in reduced dimensionality.

cr [K, r] the centroids in reduced dimensionality.
}
\description{
A function for implementing the Low-rank Canonical Correlation Analysis (LR-CCA) Algorithm.
}
\examples{
library(lol)
data <- lol.sims.rtrunk(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
model <- lol.project.lrcca(X=X, Y=Y, r=5)  # use lrcca to project into 5 dimensions
}
\author{
Jason Yim and Eric Bridgeford
}
