% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrlda.R
\name{fs.project.lrlda}
\alias{fs.project.lrlda}
\title{Low-rank Linear Discriminant Analysis (LR-LDA)}
\usage{
fs.project.lrlda(X, Y, r)
}
\arguments{
\item{X}{[n, d] the data with n samples in d dimensions.}

\item{Y}{[n] the labels of the samples.}

\item{r}{the rank of the projection.}
}
\value{
A [d, r] the projection matrix

ylabs [K] vector containing the unique, ordered class labels.

centroids [K, d] centroid matrix of the unique, ordered classes.

priors [K] vector containing prior probability for the unique, ordered classes.

Xr [n, r] the data in reduced dimensionality.

cr [K, r] the centroids in reduced dimensionality.
}
\description{
A function for implementing the LR-LDA Algorithm.
}
\author{
Richard Chen
}
