% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{lol.project.pca}
\alias{lol.project.pca}
\title{Principal Component Analysis (PCA)}
\usage{
lol.project.pca(X, r, ...)
}
\arguments{
\item{X}{[n, d] the data with n samples in d dimensions.}

\item{r}{the rank of the projection.}

\item{Y}{[n] the labels of the samples.}

\item{center=TRUE}{whether to center the data before applying PCA.}
}
\value{
A [d, r] the projection matrix from d to r dimensions.

ylabs [K] vector containing the unique, ordered class labels.

centroids [K, d] centroid matrix of the unique, ordered classes.

priors [K] vector containing prior probability for the unique, ordered classes.

Xr [n, r] the data in reduced dimensionality.

cr [K, r] the centroids in reduced dimensionality.
}
\description{
A function that performs PCA on data.
}
\examples{
library(lol)
data <- lol.sims.rtrunk(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
model <- lol.project.pca(X=X, r=2)  # use pca to project into 2 dimensions
}
\author{
Eric Bridgeford
}
