% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{fs.sims.cigar}
\alias{fs.sims.cigar}
\title{Stacked Cigar}
\usage{
fs.sims.cigar(n, d, a = 0.15, b = 4, r = FALSE)
}
\arguments{
\item{n}{the number of samples of the simulated data.}

\item{d}{the dimensionality of the simulated data.}

\item{a=0.15}{scalar for all of the mu1 but 2nd dimension.}

\item{b=4}{scalar for 2nd dimension value of mu2.}

\item{r=FALSE}{whether to apply a random rotation.}
}
\value{
X [n, d] the data as a matrix.

Y [n] the labels as a array.
}
\description{
A simulation for the stacked cigar experiment.
}
\author{
Eric Bridgeford, adapted from Joshua Vogelstein
}
