% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{fs.sims.random_rotate}
\alias{fs.sims.random_rotate}
\title{Random Rotation}
\usage{
fs.sims.random_rotate(mus, Sigmas)
}
\description{
A helper function for applying a random rotation to gaussian parameter set.
}
\author{
Eric Bridgeford
}
