% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{lol.sims.rotation}
\alias{lol.sims.rotation}
\title{Sample Random Rotation}
\usage{
lol.sims.rotation(d)
}
\description{
A helper function for estimating a random rotation matrix.
}
\author{
Eric Bridgeford
}
