% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{lol.sims.rtrunk}
\alias{lol.sims.rtrunk}
\title{Random Trunk}
\usage{
lol.sims.rtrunk(n, d, rotate = FALSE, priors = NULL, b = 4, K = 2)
}
\arguments{
\item{n}{the number of samples of the simulated data.}

\item{d}{the dimensionality of the simulated data.}

\item{rotate=FALSE}{whether to apply a random rotation to the mean and covariance. With random rotataion matrix Q, mu = Q*mu, and S = Q*S*Q.}

\item{priors=NULL}{the priors for each class. If NULL, class priors are all equal. If not null, should be |priors| = K, a length K vector for K classes.}

\item{b=4}{scalar for mu scaling.}

\item{K=2}{number of classes, should be <4.}
}
\value{
X [n, d] the data as a matrix.

Y [n] the labels as a array.

mus [d, K] the per-class means.

Sigmas [d, d, K] the per-class covariance matrices.
}
\description{
A simulation for the random trunk experiment.
}
\examples{
library(lol)
data <- lol.sims.rtrunk(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
}
\author{
Eric Bridgeford
}
