% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{fs.sims.xor2}
\alias{fs.sims.xor2}
\title{Xor Problem}
\usage{
fs.sims.xor2(D, n)
}
\arguments{
\item{D}{the dimensionality of the simulated data.}

\item{n}{the number of samples of the simulated data.}
}
\value{
X [n, d] the data as a matrix.

Y [n] the labels as a array.
}
\description{
A function to simulate from the 2-class xor problem.
}
\author{
Eric Bridgeford, adapted from Joshua Vogelstein
}
