% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdat.R
\name{fs.utils.classdat}
\alias{fs.utils.classdat}
\title{A function that performs a utility computation of prior information about the data.}
\usage{
fs.utils.classdat(X, Y)
}
\arguments{
\item{X}{[n, d] the data with n samples in d dimensions.}

\item{Y}{[n] the labels of the samples.}
}
\value{
ylabs [C] vector containing the unique, ordered class labels.

centroids [C, d] centroid matrix of the unique, ordered classes.

priors [C] vector containing prior probability for the unique, ordered classes.

n the number of samples.

d the number of dimensions.
}
\description{
A function that performs a utility computation of prior information about the data.
}
\author{
Eric Bridgeford
}
