% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval.R
\name{fs.xval.eval}
\alias{fs.xval.eval}
\title{Cross-Validation}
\usage{
fs.xval.eval(X, Y, r, alg, classifier = "lda", k = "loo")
}
\arguments{
\item{X}{[n, d] the data as n samples in d dimensions.}

\item{Y}{[n] the labels for each for each of the n samples.}

\item{r}{the number of dimensions to project the data onto.}

\item{alg=fs.project.lol}{the algorithm to use for feature selection.}

\item{classifier='lda'}{the classifier to use for assessing performance.
\itemize{
\item{'lda'}{Use the lda classifier for assessing performance.}
\item{'rf'}{Use the random forest classifier for assessing performance.}
}}

\item{k='loo'}{the cross-validated method to perform.
\itemize{
\item{'loo'}{Leave-one-out cross validation}
\item{isinteger(k)}{perform k-fold cross-validation with k as the number of folds.}
}}
}
\value{
Lhat the cross-validated error.

A [d, r] the projection matrix.

ylabs [K] vector containing the unique, ordered class labels.

centroids [K, d] centroid matrix of the unique, ordered classes.

priors [K] vector containing prior probability for the unique, ordered classes.

Xr [n, r] the data in reduced dimensionality.

cr [K, r] the centroids in reduced dimensionality.
}
\description{
A function for performing leave-one-out cross-validation for a given model.
}
\examples{
# train model and analyze with loo validation using lda classifier
library(lol)
data <- fs.sims.rtrunk(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
r=5  # embed into r=5 dimensions
xval.fit <- fs.xval.eval(X, Y, r, fs.project.lol, classifier='lda', k='loo')
}
\author{
Eric Bridgeford
}
