% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval.R
\name{fs.xval.split}
\alias{fs.xval.split}
\title{Cross-Validation Data Splitter}
\usage{
fs.xval.split(X, Y, k = "loo")
}
\arguments{
\item{X}{[n, d] an array of input data.}

\item{Y}{[s] the class labels.}

\item{k='loo'}{the cross-validated method to perform.
\itemize{
\item{'loo'}{Leave-one-out cross validation, setting k=n.}
\item{isinteger(k)}{perform k-fold cross-validation with k as the number of folds.}
}}
}
\value{
sets the cross-validation sets as a list, each element with an X.train, X.test, Y.train, and Y.test.
}
\description{
\code{sg.bern.xval_split_data} A function to split a dataset into
training and testing sets for cross validation.
}
\examples{
# prepare data for 10-fold validation
library(lol)
data <- fs.sims.rtrunk(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
sets.xval.10fold <- fs.xval.split(X, Y, k=10)
# prepare data for loo validation
sets.xval.loo <- fs.xval.split(X, Y, k='loo')

}
\author{
Eric Bridgeford
}
