% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_classifier.R
\name{lol.classify.rand}
\alias{lol.classify.rand}
\title{Random Classifier Utility}
\usage{
lol.classify.rand(X, Y, ...)
}
\arguments{
\item{X}{\code{[n, d]} the data with \code{n} samples in \code{d} dimensions.}

\item{Y}{\code{[n]} the labels of the \code{n} samples.}

\item{...}{optional args.}
}
\value{
A structure, with the following attributes:
\item{ylabs}{\code{[K]} the ylabels for each of the \code{K} unique classes, ordered.}
\item{priors}{\code{[K]} the priors for each of the \code{K} classes.}
}
\description{
A function for random classifiers.
}
\author{
Eric Bridgeford
}
