% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_optimal.R
\name{lol.project.bayes_optimal}
\alias{lol.project.bayes_optimal}
\title{Bayes Optimal}
\usage{
lol.project.bayes_optimal(X, Y, mus, Sigmas, priors, ...)
}
\arguments{
\item{X}{[n, p] the data with n samples in d dimensions.}

\item{mus}{[d, K] the true means of each class.}

\item{Sigmas}{[d, d, K] the true covariances of each class.}

\item{priors}{[K] the priors for each class.}
}
\value{
A [d, r] the projection matrix from d to r dimensions.

ylabs [K] vector containing the unique, ordered class labels.

centroids [K, d] centroid matrix of the unique, ordered classes.

priors [K] vector containing prior probability for the unique, ordered classes.

Xr [n, r] the data in reduced dimensionality.

cr [K, r] the centroids in reduced dimensionality.
}
\description{
A function for recovering the Bayes Optimal Projection, which optimizes Bayes classification.
}
\examples{
library(lol)
data <- lol.sims.rtrunk(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
model <- lol.project.bayes_optimal(X=X, mus=data$mus, S=data$Sigmas, priors=data$priors)  # obtain bayes-optimal projection of the data
}
\author{
Jason Yim and Eric Bridgeford
}
