% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{lol.sims.fat_tails}
\alias{lol.sims.fat_tails}
\title{Fat Tails Simulation}
\usage{
lol.sims.fat_tails(n, d, rotate = FALSE, f = 15, s0 = 10, rho = 0.2,
  priors = NULL)
}
\arguments{
\item{n}{the number of samples of the simulated data.}

\item{d}{the dimensionality of the simulated data.}

\item{rotate=FALSE}{whether to apply a random rotation to the mean and covariance. With random rotataion matrix Q, mu = Q*mu, and S = Q*S*Q.}

\item{priors=NULL}{the priors for each class. If NULL, class priors are all equal. If not null, should be |priors| = K, a length K vector for K classes.}

\item{f=15}{the fatness scaling of the tail. S2 = f*S1, where S1_{ij} = rho if i != j, and 1 if i == j.}

\item{s0=10}{the number of dimensions with a difference in the means. s0 should be < d.}

\item{rho=0.2}{the scaling of the off-diagonal covariance terms, should be < 1.}
}
\value{
X the data as a [n, d] matrix.

Y the labels as a [d] array.

mus [d, K] the per-class means.

Sigmas [d, d, K] the per-class covariance matrices.

priors [K] the priors for each class.
}
\description{
A function for simulating from 2 classes, one of which has fatter tails.
}
\examples{
library(lol)
data <- lol.sims.fat_tails(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
}
\author{
Eric Bridgeford
}
