% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{lol.sims.mean_diff}
\alias{lol.sims.mean_diff}
\title{Mean Difference Simulation}
\usage{
lol.sims.mean_diff(n, d, rotate = FALSE, priors = NULL, K = 2, md = 1,
  subset = c(1), offdiag = 0, s = 1)
}
\arguments{
\item{n}{the number of samples of the simulated data.}

\item{d}{the dimensionality of the simulated data.}

\item{rotate=FALSE}{whether to apply a random rotation to the mean and covariance. With random rotataion matrix \code{Q}, \code{mu = Q*mu}, and \code{S = Q*S*Q}.}

\item{priors=NULL}{the priors for each class. If \code{NULL}, class priors are all equal. If not null, should be \code{|priors| = K}, a length \code{K} vector for \code{K} classes.}

\item{K=2}{the number of classes.}

\item{md=1}{the magnitude of the difference in the means in the specified subset of dimensions.}

\item{subset=c(1)}{the dimensions to have a difference in the means. Defaults to only the first dimension. max(subset) < d.}

\item{offdiag=0}{the off-diagonal elements of the covariance matrix. Should be < 1. S_{ij} = offdiag if i != j, or 1 if i == j.}

\item{s=1}{the scaling parameter of the covariance matrix. S_{ij} = scaling*1 if i == j, or scaling*offdiag if i != j.}
}
\value{
A list of class \code{simulation} with the following:
\item{X}{\code{[n, d]} the \code{n} data points in \code{d} dimensions as a matrix.}
\item{Y}{\code{[n]} the \code{n} labels as an array.}
\item{mus}{\code{[d, K]} the \code{K} class means in \code{d} dimensions.}
\item{Sigmas}{\code{[d, d, K]} the \code{K} class covariance matrices in \code{d} dimensions.}
\item{priors}{\code{[K]} the priors for each of the \code{K} classes.}
}
\description{
A function for simulating data in which a difference in the means is present only in a subset of dimensions, and equal covariance.
}
\examples{
library(lol)
data <- lol.sims.mean_diff(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
}
\author{
Eric Bridgeford
}
