% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{lol.sims.xor2}
\alias{lol.sims.xor2}
\title{Xor Problem}
\usage{
lol.sims.xor2(n, d, priors = NULL, fall = 100)
}
\arguments{
\item{n}{the number of samples of the simulated data.}

\item{d}{the dimensionality of the simulated data.}

\item{priors=NULL}{the priors for each class. If NULL, class priors are all equal. If not null, should be |priors| = K, a length K vector for K classes.}

\item{fall=100}{the sigma for the covariance structuring. Sigma declines by ndim/fall across the variance terms.}
}
\value{
X [n, d] the data as a matrix.

Y [n] the labels as a array.

mus [d, K] the per-class means.

Sigmas [d, d, K] the per-class covariance matrices.

priors [K] the priors for each class.
}
\description{
A function to simulate from the 2-class xor problem.
}
\examples{
library(lol)
data <- lol.sims.xor2(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
}
\author{
Eric Bridgeford
}
