% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lol.utils.deltas}
\alias{lol.utils.deltas}
\title{A function that performs a utility computation of information about the differences of the classes.}
\usage{
lol.utils.deltas(centroids, priors, ...)
}
\arguments{
\item{centroids}{[d, K] centroid matrix of the unique, ordered classes.}

\item{priors}{[K] vector containing prior probability for the unique, ordered classes.}

\item{Y}{[n] the labels of the samples.}
}
\value{
deltas [d, K] the K difference vectors.
}
\description{
A function that performs a utility computation of information about the differences of the classes.
}
\author{
Eric Bridgeford
}
