% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lol.utils.info}
\alias{lol.utils.info}
\title{A function that performs basic utilities about the data.}
\usage{
lol.utils.info(X, Y, ...)
}
\arguments{
\item{X}{[n, d] the data with n samples in d dimensions.}

\item{Y}{[n] the labels of the samples.}
}
\value{
n the number of samples.

d the number of dimensions.

ylabs [K] vector containing the unique, ordered class labels.

priors [K] vector containing prior probability for the unique, ordered classes.
}
\description{
A function that performs basic utilities about the data.
}
\author{
Eric Bridgeford
}
