% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lol.utils.ohe}
\alias{lol.utils.ohe}
\title{A function for one-hot encoding categorical respose vectors.}
\usage{
lol.utils.ohe(Y)
}
\arguments{
\item{Y}{[n] a vector of the categorical resposes, with \code{K} unique categories.}
}
\value{
a list containing the following:
\item{Yh}{[n, K] the one-hot encoded Y respose variable.}
\item{ylabs}{[K] a vector of the y names corresponding to each response column.}
}
\description{
A function for one-hot encoding categorical respose vectors.
}
\author{
Eric Bridgeford
}
