% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{lol.utils.svd}
\alias{lol.utils.svd}
\title{A utility to use irlba when necessary}
\usage{
lol.utils.svd(X, nu = 0, nv = 0, t = 0.05)
}
\description{
A utility to use irlba when necessary
}
\author{
Eric Bridgeford
}
