% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval.R
\name{lol.xval.eval}
\alias{lol.xval.eval}
\title{Embedding Cross Validation}
\usage{
lol.xval.eval(X, Y, r, alg, classifier = "lda", k = "loo", ...)
}
\arguments{
\item{X}{\code{[n, d]} the data with \code{n} samples in \code{d} dimensions.}

\item{Y}{\code{[n]} the labels of the samples with \code{K} unique labels.}

\item{r}{the number of dimensions to project the data onto. Should have \code{r < d}.}

\item{alg}{the algorithm to use for embedding. Should be a function returning something of class \code{embedding}. Defaults to \code{lol.project.lol}.}

\item{classifier}{the classifier to use for assessing performance. Defaults to \code{'lda'}.
\itemize{
\item{'lda'}{ Use the lda classifier for assessing performance. \code{\link[MASS]{lda}}}
\item{'rf'}{ Use the random forest classifier for assessing performance. \code{\link[randomForest]{randomForest}}}
\item{'cent'}{ Use the nearest centroid classifier for assessing performance \code{\link{lol.classify.nearestCentroid}}}
}}

\item{k}{the cross-validated method to perform. Defaults to \code{'loo'}. \code{\link{lol.xval.split}}
\itemize{
\item{\code{'loo'}}{Leave-one-out cross validation}
\item{\code{isinteger(k)}}{ perform \code{k}-fold cross-validation with \code{k} as the number of folds.}
}}

\item{...}{optional args.}
}
\value{
Returns a list containing:
\item{Lhat}{the mean cross-validated error.}
\item{A}{\code{[d, r]} the projection matrix from \code{d} to \code{r} dimensions.}
\item{ylabs}{\code{[K]} vector containing the \code{K} unique, ordered class labels.}
\item{centroids}{\code{[K, d]} centroid matrix of \code{K} the unique, ordered classes in \code{d} dimensions.}
\item{priors}{\code{[K]} vector containing the \code{K} prior probabilities for the unique, ordered classes.}
\item{Xr}{\code{[n, r]} the \code{n} data points  in reduced dimensionality \code{r}.}
\item{cr}{\code{[K, r]} the \code{K} centroids in reduced dimensionality \code{r}.}
\item{Lhats}{the cross-validated error for each of the \code{k}-folds.}
}
\description{
A function for performing leave-one-out cross-validation for a given embedding model.
}
\examples{
# train model and analyze with loo validation using lda classifier
library(lol)
data <- lol.sims.rtrunk(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
r=5  # embed into r=5 dimensions
xval.fit <- lol.xval.eval(X, Y, r, lol.project.lol, classifier='lda', k='loo')

# train model and analyze with loo val5-fold validation using rf classifier
library(lol)
data <- lol.sims.rtrunk(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
r=5  # embed into r=5 dimensions
xval.fit <- lol.xval.eval(X, Y, r, lol.project.lol, classifier='rf', k=5)
}
\author{
Eric Bridgeford
}
