\name{print.lolMatrix}
\alias{print.lolMatrix}
\title{
print function for class lolMatrix
}
\description{
print function for class lolMatrix
}
\usage{
print.lolMatrix(x, ...)
}
\arguments{
  \item{x}{
an object of class lolMatrix
}
  \item{...}{other parameters for consistency}
}

\author{
Yinyin Yuan}



