% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol.R
\name{lol.project.lol}
\alias{lol.project.lol}
\title{Linear Optimal Low-Rank Projection (LOL)}
\usage{
lol.project.lol(X, Y, r)
}
\arguments{
\item{X}{[n, d] the data with n samples in d dimensions.}

\item{Y}{[n] the labels of the samples.}

\item{r}{the rank of the projection. Note that r >= length(unique(Y)).}
}
\value{
A [d, r] the projection matrix from d to r dimensions.

ylabs [K] vector containing the unique, ordered class labels.

centroids [K, d] centroid matrix of the unique, ordered classes.

priors [K] vector containing prior probability for the unique, ordered classes.

Xr [n, r] the data in reduced dimensionality.

cr [K, r] the centroids in reduced dimensionality.
}
\description{
A function for implementing the Linear Optimal Low-Rank Projection (LOL) Algorithm.
}
\examples{
library(lol)
data <- lol.sims.rtrunk(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
model <- lol.project.lol(X=X, Y=Y, r=5)  # use lol to project into 5 dimensions
}
\author{
Eric Bridgeford
}
