% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrcca.R
\name{fs.project.lrcca}
\alias{fs.project.lrcca}
\title{Low-rank Canonical Correlation Analysis (LR-CCA)}
\usage{
fs.project.lrcca(X, Y, r)
}
\arguments{
\item{X}{[n, p] the data with n samples in d dimensions.}

\item{Y}{[n, q] the labels of the samples.}

\item{r}{the rank of the projection.}
}
\value{
A [d, r] the projection matrix from d to r dimensions.

ylabs [K] vector containing the unique, ordered class labels.

centroids [K, d] centroid matrix of the unique, ordered classes.

priors [K] vector containing prior probability for the unique, ordered classes.

Xr [n, r] the data in reduced dimensionality.

cr [K, r] the centroids in reduced dimensionality.
}
\description{
A function for implementing the Low-rank Canonical Correlation Analysis (LR-CCA) Algorithm.
}
\author{
Jason Yim, modified by Eric Bridgeford
}
