% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{fs.project.pca}
\alias{fs.project.pca}
\title{Principal Component Analysis (PCA)}
\usage{
fs.project.pca(Xt, Y, r, center = TRUE)
}
\arguments{
\item{Y}{[n] the labels of the samples.}

\item{r}{the rank of the projection.}

\item{X}{[n, d] the data with n samples in d dimensions.}

\item{center=TRUE}{whether to center the data before applying PCA.}
}
\value{
A [d, r] the projection matrix from d to r dimensions.

ylabs [K] vector containing the unique, ordered class labels.

centroids [K, d] centroid matrix of the unique, ordered classes.

priors [K] vector containing prior probability for the unique, ordered classes.

Xr [n, r] the data in reduced dimensionality.

cr [K, r] the centroids in reduced dimensionality.
}
\description{
A function that performs PCA on data.
}
\author{
Eric Bridgeford
}
