% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{fs.sims.cigar}
\alias{fs.sims.cigar}
\title{Stacked Cigar}
\usage{
fs.sims.cigar(n, d, a = 0.15, b = 4, rotate = FALSE)
}
\arguments{
\item{n}{the number of samples of the simulated data.}

\item{d}{the dimensionality of the simulated data.}

\item{a=0.15}{scalar for all of the mu1 but 2nd dimension.}

\item{b=4}{scalar for 2nd dimension value of mu2.}

\item{rotate=FALSE}{whether to apply a random rotation.}
}
\value{
X [n, d] the data as a matrix.

Y [n] the labels as a array.

mus [d, K] the per-class means.

Sigmas [d, d, K] the per-class covariance matrices.
}
\description{
A simulation for the stacked cigar experiment.
}
\examples{
library(fselect)
data <- fs.sims.cigar(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
}
\author{
Eric Bridgeford, adapted from Joshua Vogelstein
}
