% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{fs.sims.rotation}
\alias{fs.sims.rotation}
\title{Sample Random Rotation}
\usage{
fs.sims.rotation(d)
}
\description{
A helper function for estimating a random rotation matrix.
}
\author{
Eric Bridgeford
}
