% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{fs.sims.rtrunk}
\alias{fs.sims.rtrunk}
\title{Random Trunk}
\usage{
fs.sims.rtrunk(n, d, b = 4, r = FALSE, C = 2)
}
\arguments{
\item{n}{the number of samples of the simulated data.}

\item{d}{the dimensionality of the simulated data.}

\item{b=4}{scalar for mu scaling.}

\item{r=FALSE}{whether to apply a random rotation.}

\item{C=2}{number of classes, should be <4.}
}
\value{
X [n, d] the data as a matrix.

Y [n] the labels as a array.
}
\description{
A simulation for the random trunk experiment.
}
\author{
Eric Bridgeford, adapted from Joshua Vogelstein
}
