% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{fs.sims.sim_gmm}
\alias{fs.sims.sim_gmm}
\title{GMM Simulate}
\usage{
fs.sims.sim_gmm(mus, Sigmas, n, priors = NULL)
}
\arguments{
\item{mus}{[d, C] the mus for each class.}

\item{Sigmas}{[d,d,C] the Sigmas for each class.}

\item{n}{the of examples.}

\item{priors=NULL}{the priors for each class. If NULL, class priors are all equal. If not null, should be |priors| = K, a length K vector.}
}
\value{
X [n, d] the simulated data.

Y [n] the labels for each data point.
}
\description{
A helper function for simulating from Gaussian Mixture.
}
\author{
Eric Bridgeford
}
