% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{fs.sims.toep}
\alias{fs.sims.toep}
\title{Toeplitz Simulation}
\usage{
fs.sims.toep(n, d, D1 = 10, r = FALSE, b = 0.4, rho = 0.5)
}
\arguments{
\item{n}{the number of samples of the simulated data.}

\item{d}{the dimensionality of the simulated data.}

\item{D1=10}{the dimensionality for the non-equal covariance terms}

\item{r=FALSE}{whether to apply a random rotation.}

\item{rho=0.5}{the scaling of the covariance terms, should be < 1.}
}
\value{
X [n, d] the data as a matrix.

Y [n] the labels as a array.
}
\description{
A function for simulating data in which the covariance is a non-symmetric toeplitz matrix.
}
\author{
Eric Bridgeford, adapted from Joshua Vogelstein
}
