% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{fs.sims.toep}
\alias{fs.sims.toep}
\title{Toeplitz Simulation}
\usage{
fs.sims.toep(n, d, D1 = 10, rotate = FALSE, b = 0.4, rho = 0.5)
}
\arguments{
\item{n}{the number of samples of the simulated data.}

\item{d}{the dimensionality of the simulated data.}

\item{D1=10}{the dimensionality for the non-equal covariance terms.}

\item{rotate=FALSE}{whether to apply a random rotation.}

\item{b=0.4}{a scaling parameter for the means.}

\item{rho=0.5}{the scaling of the covariance terms, should be < 1.}
}
\value{
X [n, d] the data as a matrix.

Y [n] the labels as a array.
}
\description{
A function for simulating data in which the covariance is a non-symmetric toeplitz matrix.
}
\examples{
library(fselect)
data <- fs.sims.toep(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
}
\author{
Eric Bridgeford, adapted from Joshua Vogelstein
}
