% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{fs.sims.xor2}
\alias{fs.sims.xor2}
\title{Xor Problem}
\usage{
fs.sims.xor2(n, d, fall = 100)
}
\arguments{
\item{n}{the number of samples of the simulated data.}

\item{d}{the dimensionality of the simulated data.}

\item{fall=20}{the sigma for the covariance structuring.}
}
\value{
X [n, d] the data as a matrix.

Y [n] the labels as a array.
}
\description{
A function to simulate from the 2-class xor problem.
}
\author{
Eric Bridgeford, adapted from Joshua Vogelstein
}
