% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{lol.sims.mean_diff}
\alias{lol.sims.mean_diff}
\title{Mean Difference Simulation}
\usage{
lol.sims.mean_diff(n, d, rotate = FALSE, priors = NULL, K = 2, md = 1,
  subset = c(1), offdiag = 0, s = 1)
}
\arguments{
\item{n}{the number of samples of the simulated data.}

\item{d}{the dimensionality of the simulated data.}

\item{rotate=FALSE}{whether to apply a random rotation to the mean and covariance. With random rotataion matrix Q, mu = Q*mu, and S = Q*S*Q.}

\item{priors=NULL}{the priors for each class. If NULL, class priors are all equal. If not null, should be |priors| = K, a length K vector for K classes.}

\item{K=2}{the number of classes.}

\item{md=1}{the magnitude of the difference in the means in the specified subset of dimensions.}

\item{subset=c(1)}{the dimensions to have a difference in the means. Defaults to only the first dimension. max(subset) < d.}

\item{offdiag=0}{the off-diagonal elements of the covariance matrix. Should be < 1. S_{ij} = offdiag if i != j, or 1 if i == j.}

\item{s=1}{the scaling parameter of the covariance matrix. S_{ij} = scaling*1 if i == j, or scaling*offdiag if i != j.}
}
\value{
X [n, d] the data as a matrix.

Y [n] the labels as a array.

mus [d, K] the per-class means.

Sigmas [d, d, K] the per-class covariance matrices.
}
\description{
A function for simulating data in which a difference in the means is present only in a subset of dimensions, and equal covariance.
}
