% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{lol.sims.sim_gmm}
\alias{lol.sims.sim_gmm}
\title{GMM Simulate}
\usage{
lol.sims.sim_gmm(mus, Sigmas, n, priors)
}
\arguments{
\item{mus}{[d, K] the mus for each class.}

\item{Sigmas}{[d,d,K] the Sigmas for each class.}

\item{n}{the of examples.}

\item{priors}{the priors for each class.}
}
\value{
X [n, d] the simulated data.

Y [n] the labels for each data point.

priors [K] the priors for each class.
}
\description{
A helper function for simulating from Gaussian Mixture.
}
\author{
Eric Bridgeford
}
