% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{lol.sims.sim_gmm}
\alias{lol.sims.sim_gmm}
\title{GMM Simulate}
\usage{
lol.sims.sim_gmm(mus, Sigmas, n, priors)
}
\arguments{
\item{mus}{\code{[d, K]} the mus for each class.}

\item{Sigmas}{\code{[d,d,K]} the Sigmas for each class.}

\item{n}{the number of examples.}

\item{priors}{\code{K} the priors for each class.}
}
\value{
A list with the following:
\item{X}{\code{[n, d]} the simulated data.}
\item{Y}{\code{[n]} the labels for each data point.}
\item{priors}{\code{[K]} the priors for each class.}
}
\description{
A helper function for simulating from Gaussian Mixture.
}
\author{
Eric Bridgeford
}
