% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{lol.sims.toep}
\alias{lol.sims.toep}
\title{Toeplitz Simulation}
\usage{
lol.sims.toep(n, d, rotate = FALSE, priors = NULL, D1 = 10, b = 0.4,
  rho = 0.5)
}
\arguments{
\item{n}{the number of samples of the simulated data.}

\item{d}{the dimensionality of the simulated data.}

\item{D1=10}{the dimensionality for the non-equal covariance terms.}

\item{rotate=FALSE}{whether to apply a random rotation to the mean and covariance. With random rotataion matrix Q, mu = Q*mu, and S = Q*S*Q.}

\item{priors=NULL}{the priors for each class. If NULL, class priors are all equal. If not null, should be |priors| = K, a length K vector for K classes.}

\item{b=0.4}{a scaling parameter for the means.}

\item{rho=0.5}{the scaling of the covariance terms, should be < 1.}
}
\value{
X [n, d] the data as a matrix.

Y [n] the labels as a array.

mus [d, K] the per-class means.

Sigmas [d, d, K] the per-class covariance matrices.

priors [K] the priors for each class.
}
\description{
A function for simulating data in which the covariance is a non-symmetric toeplitz matrix.
}
\examples{
library(lol)
data <- lol.sims.toep(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
}
\author{
Eric Bridgeford
}
