% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lol.utils.deltas}
\alias{lol.utils.deltas}
\title{A function that performs a utility computation of information about the differences of the classes.}
\usage{
lol.utils.deltas(centroids, priors, ...)
}
\arguments{
\item{centroids}{\code{[d, K]} centroid matrix of the unique, ordered classes.}

\item{priors}{\code{[K]} vector containing prior probability for the unique, ordered classes.}

\item{...}{optional args.}
}
\value{
deltas \code{[d, K]} the K difference vectors.
}
\description{
A function that performs a utility computation of information about the differences of the classes.
}
\author{
Eric Bridgeford
}
