% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{lol.utils.svd}
\alias{lol.utils.svd}
\title{A utility to use irlba when necessary}
\usage{
lol.utils.svd(X, nu = 0, nv = 0, t = 0.05)
}
\arguments{
\item{X}{the data to compute the svd of.}

\item{nu}{the number of left singular vectors to retain.}

\item{nv}{the number of right singular vectors to retain.}

\item{t}{the threshold of percent of singular vals/vecs to use irlba.}
}
\value{
the svd of X.
}
\description{
A utility to use irlba when necessary
}
\author{
Eric Bridgeford
}
