% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval.R
\name{lol.xval.split}
\alias{lol.xval.split}
\title{Cross-Validation Data Splitter}
\usage{
lol.xval.split(X, Y, k = "loo", reverse = FALSE, ...)
}
\arguments{
\item{X}{\code{[n, d]} the data with \code{n} samples in \code{d} dimensions.}

\item{Y}{\code{[n]} the labels of the samples with \code{K} unique labels.}

\item{k}{the cross-validated method to perform. Defaults to \code{'loo'}.
\itemize{
\item{if \code{k == round(k)}, performed k-fold cross-validation.}
\item{if \code{k == 'loo'}, performs leave-one-out cross-validation.}
}}

\item{reverse}{whether to flip the training and testing partitions. Defaults to \code{FALSE}.
\itemize{
\item{if \code{isTRUE(reverse)}, flips the training and testing partitions, such that cross-validation will be performed with 1 training fold and k-1 testing folds.}
\item{if \code{!isTRUE(reverse)}, standard cross-validation, with k-1 training folds and 1 testing fold.}
}}

\item{...}{optional args.}
}
\value{
sets the cross-validation sets as an object of class \code{"XV"}. Each element of the list contains the following items:
\item{\code{X.train}}{the training data as a \code{[n - n/k, d]} array.}
\item{\code{Y.train}}{the training labels as a \code{[n - n/k]} vector.}
\item{\code{X.test}}{the testing data as a \code{[n/k, d]} array.}
\item{\code{Y.test}}{the testing labels as a \code{[n/k]} vector.}
}
\description{
A function to split a dataset into
training and testing sets for cross validation.
}
\examples{
# prepare data for 10-fold validation
library(lolR)
data <- lol.sims.rtrunk(n=200, d=30)  # 200 examples of 30 dimensions
X <- data$X; Y <- data$Y
sets.xval.10fold <- lol.xval.split(X, Y, k=10)

# prepare data for loo validation
sets.xval.loo <- lol.xval.split(X, Y, k='loo')

}
\author{
Eric Bridgeford
}
